debug: Abstract out the destination of debug messages
Don't worry about where debug messages end up. Instead, log() a
message when there is something to log, or send a dbg() message when
appropriate. This saves the very tedious typing of:

D(fprintf, stderr, "msg") ==> reink_dgb("msg");

This is only implemented for reink, but d4lib is not modified.