util: Add utility function for redirecting log output
Add a backed funtion to help with redirecting debug output. Note
that while the messages are still sent to stderr, redirecting them to
a file, network, or any type of stream can now be done with minimal
modifications to reink_log. It's also possible to preprend a message
to each line and so on.

The scope of this change, however, is to enable this flexibility and
improve readability. It is not to implement more output streams.