reink: Add support for binary status reports 'ST2'
Newer printers will reply to the status request with a binary packet.
Since this packet contains the text "ST2", as opposed to "ST" in the
legacy packets, refer to it as v2 status report throughout the code.

It's fairly simple to distinguish a v2 report from a v1 report, and
we just call the appropriate handler based on the report we recieve.

For the v2 report, auto-identification of inks and readink ink levels
is currently implemented.