reink: Implement an object-oriented printer model
Passing around fds and socket_ids, as well as indices into global was
extremely fun, but there is a more elegant way. From the file
descriptors, to the IEEE1284.4 sockets, to the printer database model,
and even the details of d4lib, there is a lot that can be abstracted
away.

A printer has a set of fixed characteristics, such as the model name,
EEPROM layout. On the other hand is has dynamic characterictics, such
the physical connection. Several data streams (.4 sockets) can flow
over this connection.

The approach here tries to model that, with the intent of greatly
reducing the burden of argument passing between function calls. The
specifics are explained in the comments.